Revert "Move frontend package types out of typescript-typings since this package will be used in both node and browser contexts"
This reverts commit 74483631d09395de248fe97034c059a2a8bb2097.
