+++ packages/typescript-typings/package.json
@@ -25,9 +25,9 @@
     },
     "homepage": "https://github.com/0xProject/0x-monorepo/packages/typescript-typings#readme",
     "dependencies": {
-        "@types/bn.js": "^4.11.0",
+        "ethereum-types": "^1.0.0",
         "bignumber.js": "~4.1.0",
-        "ethereum-types": "^1.0.0"
+        "@types/bn.js": "^4.11.0"
     },
     "devDependencies": {
         "@0xproject/monorepo-scripts": "^1.0.0",
