+++ packages/typescript-typings/types/react-highlight/index.d.ts
@@ -1,0 +1,1 @@
+declare module 'react-highlight';
