+++ packages/typescript-typings/types/react-tooltip/index.d.ts
@@ -1,0 +1,1 @@
+declare module 'react-tooltip';
